Support region parameter for aws resources
As of version 6.0 of the AWS provider, it is possible to use a single provider instance to deploy resources to multiple AWS regions. This change adds support for a `region` input variable. If this is not provided, the module will continue to create resources in the region configured by the provider.